refactor: UserInputValidator 제거
